refactor: conform sources to Principle VI
Split each dual-kind feed into thin FootballData/Sample/basketball classes in
_stats and _odds over a shared _common subpackage, removing the wholesale file
duplication. Rename the source contract and schema methods verb-first, collapse
the private-helper docstrings, and route the shared IDENTITY_COLS and the
basketball constants/logic through core and _common/_basketball.

Read the odds-api key from a named environment variable instead of taking it as
an argument, so the secret is never passed or pickled. Privatise map_odds_names
and BaseSchema, document the public resolver functions, and drop MATCH_COLS from
the sources surface (it lives in core).

BREAKING CHANGE: source methods index_items/catalogue/available_params/required_items/fixtures_items are now list_index_items/read_catalogue/list_available_params/list_required_items/list_fixtures_items; schema methods snapshot_cols/col_metadata/odds_cols are now list_snapshot_cols/get_col_metadata/list_odds_cols; and OddsApi(key=...) is now OddsApi(key_env=...).
