+++ src/sportsbet/cli/_data.py
@@ -38,7 +38,7 @@ def params(**selection: object) -> None:
         if loader is None:
             return
         stats_source, *_ = loader.sources_
-        available = stats_source.available_params()
+        available = stats_source.list_available_params()
         cols = list({param for params in available for param in params})
         frame = pd.DataFrame({col: [params.get(col, '-') for params in available] for col in cols})
         print_console([frame], ['Available parameters'], index=False)
