+++ src/sportsbet/mcp/_server.py
@@ -107,7 +107,7 @@ def _extraction(
 def _available_params(selection: Selection) -> list[dict]:
     """Return what can be selected."""
     stats_source, *_ = build_dataloader(**selection).sources_
-    return stats_source.available_params()
+    return stats_source.list_available_params()
 
 
 def _odds_types(selection: Selection) -> list[str]:
