+++ src/sportsbet/sources/_common/__init__.py
@@ -1,0 +1,1 @@
+"""Define the code shared by a feed's statistics and odds sources."""
