+++ src/sportsbet/sources/_stats/_nba.py
@@ -14,18 +14,12 @@ import pandas as pd
 
 from ...core import ParamGrid
 from .._base import BaseStatsSource, RawItem, RawPayload
-from .._utils import derive_market_outcomes
+from .._common._basketball import DIVISION, SEASONS_KEY, _snapshots
 
 SEASONS_URL = 'https://sports.core.api.espn.com/v2/sports/basketball/leagues/nba/seasons?limit=100'
 GAMES_URL = 'https://site.api.espn.com/apis/site/v2/sports/basketball/nba/scoreboard?dates={start}-{end}&limit=1000'
 
 LEAGUE = 'NBA'
-DIVISION = 1
-MARKETS = ['home_win', 'away_win']
-SEASONS_KEY = 'seasons'
-ROLLING_GAMES = 3
-FEATURES = ['points_for', 'points_against', 'wins']
-IDENTITY = ['date', 'league', 'division', 'year', 'home_team', 'away_team']
 
 PRESEASON = 1
 EXHIBITION = 'ALLSTAR'
@@ -33,15 +27,7 @@ MONTHS = [(-1, month) for month in (9, 10, 11, 12)] + [(0, month) for month in (
 
 
 def _wanted(event: dict[str, Any]) -> bool:
-    """Return whether an event is a game of the competition rather than an exhibition.
-
-    The two labels an event carries each lie on their own. The all-star games are filed under the regular season, and
-    their teams are inventions, so keeping the regular season keeps them. The play-off rounds are not filed as standard
-    games, so keeping the standard games drops every one of them.
-
-    So it keeps an event only when both labels are recognized, which keeps the pairing with the odds sound: a dropped
-    game shows as a visible gap, while an invented team in the roster would corrupt the whole competition's pairing.
-    """
+    """Return whether an event is a game of the competition rather than an exhibition."""
     competitions = event.get('competitions') or [{}]
     season_type = event.get('season', {}).get('type')
     competition_type = competitions[0].get('type', {}).get('abbreviation')
@@ -49,12 +35,7 @@ def _wanted(event: dict[str, Any]) -> bool:
 
 
 def _games(content: bytes, year: int) -> pd.DataFrame:
-    """Return the games of a month, as the API publishes them.
-
-    The tip-off is the instant the API gives, which it gives in UTC. Whether a game was played is the flag it carries on
-    that game: a finished season still holds games that were postponed and never made up, and those have no result to
-    learn from.
-    """
+    """Return the games of a month, with the API's UTC tip-off and its played flag."""
     events = json.loads(content).get('events', [])
     records = []
     for event in events:
@@ -85,76 +66,6 @@ def _games(content: bytes, year: int) -> pd.DataFrame:
     return frame
 
 
-def _form(games: pd.DataFrame) -> pd.DataFrame:
-    """Return what each team had done before each of its games.
-
-    The upcoming games are already part of the frame, so one of them carries the form of the games before it. Every
-    average is shifted by one, so a game sees only the games before it.
-
-    Points scored, points conceded, and the wins that follow from them. The feed carries a score line, so that is what
-    the form is built from.
-    """
-    played = games['home_points'].ge(0) & games['away_points'].ge(0)
-    sides = [
-        pd.DataFrame(
-            {
-                'team': games[f'{side}_team'],
-                'date': games['date'],
-                'points_for': games[f'{side}_points'].where(played),
-                'points_against': games[f'{other}_points'].where(played),
-                'wins': (games[f'{side}_points'] > games[f'{other}_points']).where(played).astype(float),
-            },
-        )
-        for side, other in (('home', 'away'), ('away', 'home'))
-    ]
-    form = pd.concat(sides).set_index(['team', 'date']).sort_index()
-
-    averages = [f'{col}_avg' for col in FEATURES]
-    latest = [f'{col}_latest_avg' for col in FEATURES]
-    form[averages] = form.groupby('team')[FEATURES].expanding().mean().to_numpy()
-    form[averages] = form.groupby('team')[averages].shift(1)
-    form[latest] = form.groupby('team')[FEATURES].rolling(window=ROLLING_GAMES, min_periods=1).mean().to_numpy()
-    form[latest] = form.groupby('team')[latest].shift(1)
-    return form.drop(columns=FEATURES).reset_index()
-
-
-def _snapshots(games: pd.DataFrame) -> pd.DataFrame:
-    """Return the long snapshots of a season.
-
-    A game that has not been played gets only its pre-play snapshot, so it becomes a fixture.
-    """
-    if games.empty:
-        return games
-    form = _form(games)
-    feature_cols = [col for col in form.columns if col.endswith('avg')]
-    preplay = games[IDENTITY].copy()
-    for side in ('home', 'away'):
-        sided = [f'{side}_{col}' for col in feature_cols]
-        side_form = form.rename(columns=dict(zip(feature_cols, sided, strict=True)))
-        preplay = preplay.merge(
-            side_form[['team', 'date', *sided]],
-            left_on=['date', f'{side}_team'],
-            right_on=['date', 'team'],
-            how='left',
-        ).drop(columns='team')
-    preplay = preplay.assign(event_status='preplay', event_time=0)
-
-    played = games['home_points'].ge(0) & games['away_points'].ge(0)
-    postplay = games.loc[played, IDENTITY].assign(
-        home_points=games.loc[played, 'home_points'].astype(int),
-        away_points=games.loc[played, 'away_points'].astype(int),
-        event_status='postplay',
-        event_time=0,
-    )
-    outcomes = derive_market_outcomes(postplay['home_points'], postplay['away_points'], MARKETS)
-    postplay = pd.concat([postplay, outcomes], axis=1)
-
-    snapshots = pd.concat([preplay, postplay], ignore_index=True)
-    sided = [f'{side}_{col}' for side in ('home', 'away') for col in feature_cols]
-    order = ['event_status', 'event_time', *IDENTITY, 'home_points', 'away_points', *MARKETS, *sided]
-    return snapshots.reindex(columns=[col for col in order if col in snapshots.columns])
-
-
 class NBAStats(BaseStatsSource):
     """The statistics of the NBA, as ESPN publishes them.
 
@@ -179,7 +90,7 @@ class NBAStats(BaseStatsSource):
         >>> dataloader = DataLoader(
         ...     param_grid={'league': ['NBA'], 'year': [2026]},
         ...     stats=source,
-        ...     odds=OddsApi(key='...', markets=['h2h']),
+        ...     odds=OddsApi(key_env='ODDS_API_KEY', markets=['h2h']),
         ... )
         >>> dataloader.sport_
         'basketball'
@@ -192,16 +103,12 @@ class NBAStats(BaseStatsSource):
     sport: ClassVar[str | None] = 'basketball'
     name: ClassVar[str] = 'nba'
 
-    def index_items(self: Self, selection: ParamGrid | None = None) -> list[RawItem]:
+    def list_index_items(self: Self, selection: ParamGrid | None = None) -> list[RawItem]:
         """Return the seasons the competition publishes, which is one free request whatever is selected."""
         return [RawItem(source=self.name, key=SEASONS_KEY, url=SEASONS_URL)]
 
-    def catalogue(self: Self, payloads: list[RawPayload]) -> list[dict]:
-        """Return the seasons the competition publishes.
-
-        A season is named by the year it ends in, as everywhere else in the library, and that is already the year the
-        feed gives. There is nothing to convert here, unlike the EuroLeague, whose `E2024` is the 2024-25 season.
-        """
+    def read_catalogue(self: Self, payloads: list[RawPayload]) -> list[dict]:
+        """Return the seasons the competition publishes, each named by the year it ends in."""
         if not payloads:
             return []
         seasons = json.loads(payloads[0].content).get('items', [])
@@ -211,13 +118,8 @@ class NBAStats(BaseStatsSource):
             key=lambda params: params['year'],
         )
 
-    def required_items(self: Self, params: list[dict], schedule: pd.DataFrame | None = None) -> list[RawItem]:
-        """Return one item per month of each selected season.
-
-        The feed returns at most a thousand games and says nothing when it has more, so asking for a whole season, which
-        is about fourteen hundred, quietly loses a quarter of it. A month holds two hundred and forty at its busiest, so
-        a month is an answer the feed can give in full.
-        """
+    def list_required_items(self: Self, params: list[dict], schedule: pd.DataFrame | None = None) -> list[RawItem]:
+        """Return one item per month of each selected season, since the feed caps a request at a thousand games."""
         items = []
         for param in params:
             if param['league'] != LEAGUE:
