+++ tests/dataloaders/test_basketball.py
@@ -50,7 +50,7 @@ def test_the_sport_is_the_source_s_and_not_the_dataloader_s():
     """Test a feed of basketball is a feed of basketball, whatever it is handed to."""
     assert EuroLeagueStats.sport == 'basketball'
     assert NBAStats.sport == 'basketball'
-    assert DataLoader(stats=NBAStats(), odds=OddsApi(key='k')).sport_ == 'basketball'
+    assert DataLoader(stats=NBAStats(), odds=OddsApi(key_env='ODDS_API_KEY')).sport_ == 'basketball'
 
 
 def test_statistics_and_odds_of_different_sports_are_refused():
